+++ lib/maze/api/appium/file_manager.rb
@@ -37,7 +37,7 @@ module Maze
         rescue Selenium::WebDriver::Error::ServerError => e
           $logger.error "Error writing file to device: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver(e.message)
+          fail_driver(e)
           raise e
         end
 
@@ -75,7 +75,7 @@ module Maze
         rescue Selenium::WebDriver::Error::ServerError => e
           $logger.error "Error reading file from device: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver(e.message)
+          fail_driver(e)
           raise e
         end
 
@@ -105,7 +105,7 @@ module Maze
         rescue Selenium::WebDriver::Error::ServerError => e
           $logger.error "Error reading folder from device: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver(e.message)
+          fail_driver(e)
           raise e
         end
       end